requirements.txt
requests
beautifulsoup4
feedparser
openai
python-dotenv
flask
flask-cors
stripe
resend
